# modules/auth/controller/auth_controller.go
package controller

import (
	"net/http"

	"github.com/Mobilizes/materi-be-alpro/modules/auth/dto"
	"github.com/Mobilizes/materi-be-alpro/modules/auth/service"
	"github.com/Mobilizes/materi-be-alpro/modules/auth/validation"
	"github.com/Mobilizes/materi-be-alpro/pkg/utils"
	"github.com/gin-gonic/gin"
)

type AuthController struct {
	authService *service.AuthService
}

func NewAuthController(authService *service.AuthService) *AuthController {
	return &AuthController{authService: authService}
}

// @Summary User Login
// @Description Authenticate user with email and password
// @Tags Auth
// @Accept json
// @Produce json
// @Param login body dto.LoginRequest true "Login Credentials"
// @Success 200 {object} utils.SwaggerSuccessResponse{data=dto.TokenResponse} "Login Successful"
// @Failure 400 {object} utils.SwaggerErrorResponse "Invalid Request"
// @Failure 401 {object} utils.SwaggerErrorResponse "Unauthorized"
// @Router /auth/login [post]
func (ctrl *AuthController) Login(c *gin.Context) {
	req, err := validation.ValidateLogin(c)
	if err != nil {
		utils.ErrorResponse(c, http.StatusBadRequest, err.Error())
		return
	}

	token, err := ctrl.authService.Login(req)
	if err != nil {
		utils.ErrorResponse(c, http.StatusUnauthorized, err.Error())
		return
	}

	utils.SuccessResponse(c, http.StatusOK, "Login berhasil", dto.TokenResponse{Token: token})
}